test: cli CI fails clearer and faster
